services/client/requirements.txt
fastapi
uvicorn
aio-pika
python-decouple
websockets
